# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Wishlist API","version":"1.0.0"},"paths":{"\/api\/register\/":{"post":{"tags":["auth"],"summary":"Registra un nuovo utente","operationId":"App\\Http\\Controllers\\AuthController::register","parameters":[{"name":"name","in":"query","description":"username","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","description":"email (univoca)","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"password","required":true,"schema":{"type":"string"}}],"responses":{"201":{"description":"Utente creato con successo","content":{"application\/json":{"schema":{}}}},"409":{"description":"Errore creazione utente \/ Eccezione"}}}},"\/api\/login\/":{"post":{"tags":["auth"],"summary":"Get a JWT via given credentials.","operationId":"App\\Http\\Controllers\\AuthController::login","parameters":[{"name":"email","in":"query","description":"email","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","description":"password","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Autenticazione effettuata","content":{"application\/json":{"schema":{"properties":{"":{"title":"token","description":"JWT token","type":"string"}},"type":"object"}}}}}}},"\/api\/v1\/products\/":{"post":{"tags":["products"],"summary":"Aggiunge prodotto","operationId":"App\\Http\\Controllers\\V1\\ProductController::add","parameters":[{"name":"name","in":"query","description":"Nome del prodotto","required":true,"schema":{"type":"string"}},{"name":"sku","in":"query","description":"Codice univoco del prodotto","required":true,"schema":{"type":"string"}},{"name":"price","in":"query","description":"Prezzo","required":true,"schema":{"type":"float"}},{"name":"description","in":"query","description":"Descrizione prodotto","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Prodotto creato con successo","content":{"application\/json":{"schema":{"properties":{"":{"title":"id","description":"Id prodotto creato","type":"integer"}},"type":"object"}}}},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]}},"\/api\/v1\/products\/{id}":{"delete":{"tags":["products"],"summary":"Deletes by index key","operationId":"App\\Http\\Controllers\\V1\\ProductController::remove","parameters":[{"name":"id","in":"path","description":"Id prodotto da eliminare","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"description":"Prodotto cancellato con successo"},"404":{"description":"Prodotto non trovato"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]}},"\/api\/v1\/wishlists\/{id}":{"get":{"tags":["wishlist"],"operationId":"App\\Http\\Controllers\\V1\\WishlistController::fetch","parameters":[{"name":"id","in":"path","description":"Id wishlist da recuperare","required":false,"schema":{"type":"int"}}],"responses":{"200":{"description":"Restituisce la wishlist richiesta o tutte le wishlist dell'utente","content":{"application\/json":{"schema":{}}}},"404":{"description":"La wishlist non esiste, o l'id non corrisponde ad una wishlist valida"},"500":{"description":"Errore"}},"security":[{"bearer":[]}]},"put":{"tags":["wishlist"],"summary":"Modifica la wishlist","operationId":"App\\Http\\Controllers\\V1\\WishlistController::edit","parameters":[{"name":"id","in":"path","description":"Id wishlist da modificare","required":true,"schema":{"type":"integer"}},{"name":"name","in":"query","description":"Nome da assegnare alla wishlist","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Wishlist modificata con successo","content":{"application\/json":{"schema":{"properties":{"":{"title":"id","description":"Id prodotto modificato","type":"integer"}},"type":"object"}}}},"404":{"description":"Wishlist non trovata"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]},"delete":{"tags":["wishlist"],"summary":"Deletes by index key","operationId":"App\\Http\\Controllers\\V1\\WishlistController::remove","parameters":[{"name":"id","in":"path","description":"Id wishlist da eliminare","required":true,"schema":{"type":"integer"}}],"responses":{"204":{"description":"Wishlist cancellata con successo"},"404":{"description":"Wishlist non trovata"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]}},"\/api\/v1\/wishlists\/":{"post":{"tags":["wishlist"],"operationId":"App\\Http\\Controllers\\V1\\WishlistController::add","parameters":[{"name":"name","in":"query","description":"Nome da assegnare alla wishlist","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Wishlist creata con successo","content":{"application\/json":{"schema":{"properties":{"":{"title":"id","description":"Id wishlist creata","type":"integer"}},"type":"object"}}}},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]}},"\/api\/v1\/wishlists\/{id}\/products":{"get":{"tags":["wishlist-products"],"summary":"Recupera tutti i prodotti relativi all wishlist indicata","operationId":"App\\Http\\Controllers\\V1\\WishlistController::fetchProducts","parameters":[{"name":"id","in":"path","description":"Id wishlist da cui estrarre i prodotti","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Elenco prodotti","content":{"application\/json":{"schema":{}}}},"404":{"description":"Wishlist non trovata"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]},"post":{"tags":["wishlist-products"],"summary":"Aggiunge il prodotto $pid alla lista $wid","operationId":"App\\Http\\Controllers\\V1\\WishlistController::addProducts","parameters":[{"name":"id","in":"path","description":"Id wishlist da cui estrarre i prodotti","required":true,"schema":{"type":"integer"}},{"name":"pid","in":"query","description":"Id prodotto da aggiungere","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Elenco prodotti","content":{"application\/json":{"schema":{}}}},"301":{"description":"Prodotto gi\u00e0 presente"},"404":{"description":"Wishlist o prodotto non trovati"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]},"delete":{"tags":["wishlist-products"],"summary":"ELimina il prodotto passato dalla lista $wid","operationId":"App\\Http\\Controllers\\V1\\WishlistController::removeProducts","parameters":[{"name":"id","in":"path","description":"Id wishlist da cui estrarre i prodotti","required":true,"schema":{"type":"integer"}},{"name":"pid","in":"query","description":"Id prodotto da rimuovere","required":true,"schema":{"type":"integer"}}],"responses":{"404":{"description":"Wishlist o prodotto trovati"},"500":{"description":"Errore \/ Eccezione"}},"security":[{"bearer":[]}]}}},"components":{"securitySchemes":{"bearer":{"type":"http","description":"Authorization token obtained from logging in.","name":"Authorization","in":"header","scheme":"bearer"}}}}